feat(zfs): Ensure same pool is not used by two benches
